--- .dockerignore
@@ -1,4 +1,0 @@
-node_modules/
-dist/
-lib/
-__tests__/
