+++ .node-version
@@ -1,0 +1,1 @@
+24.9.0
