+++ package.json
@@ -21,23 +21,25 @@
     "url": "https://francoisbest.com"
   },
   "type": "module",
-  "main": "dist/main.js",
+  "main": "src/main.ts",
   "scripts": {
-    "build": "tsup ./src/main.ts --format esm",
-    "test": "vitest"
+    "dev": "node --env-file=.env.local --watch ./src/main.ts",
+    "test": "vitest --typecheck"
+  },
+  "engines": {
+    "node": "^24.9.0"
   },
   "dependencies": {
-    "@actions/core": "^1.10.1",
+    "@actions/core": "^1.11.1",
     "@actions/exec": "^1.1.1",
-    "clever-tools": "^3.5.0"
+    "clever-tools": "^4.1.0"
   },
   "devDependencies": {
-    "@types/node": "^20.12.6",
-    "@vitest/spy": "^1.4.0",
+    "@types/node": "^24.6.2",
+    "@vitest/spy": "^3.2.4",
     "prettier": "^3.2.5",
-    "tsup": "^8.0.2",
-    "typescript": "^5.4.4",
-    "vitest": "^1.4.0"
+    "typescript": "^5.9.3",
+    "vitest": "^3.2.4"
   },
   "prettier": {
     "printWidth": 80,
