+++ tsconfig.json
@@ -17,11 +17,11 @@
     "declaration": true,
     "declarationMap": true,
     "verbatimModuleSyntax": true,
+    "isolatedDeclarations": true,
     "moduleDetection": "force",
 
     "downlevelIteration": true,
     // Interop
-    "allowJs": true,
     "isolatedModules": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
@@ -31,6 +31,10 @@
     "incremental": true,
     "tsBuildInfoFile": ".tsbuildinfo"
   },
-  "include": ["**/*.ts", "__tests__/processArguments.test.cts", "__tests__/main.test.ts.disabled"],
+  "include": [
+    "**/*.ts",
+    "__tests__/processArguments.test.cts",
+    "__tests__/main.test.ts.disabled"
+  ],
   "exclude": ["node_modules"]
 }
