+++ vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig, type ViteUserConfig } from 'vitest/config'
+
+const config: ViteUserConfig = defineConfig({
+  test: {
+    environment: 'node'
+  }
+})
+
+export default config
